Redevelop Timeseries with tests
Turns out there is a TON of odd/difficult behavior around time.  Among
the details accounted for below...

MST != MST7MDT
---------------

Time zones are specific and peculiar. Time can identify a zone offset as well
as a zone abbreviation like 'MST' or 'MDT' but it appears to be unable to
return the actual zone name like 'MST7MDT' or 'US/Mountain' or
'America/Denver'.  Moreover, the abbreviations cannot always be mapped back to
the full zone name.  'CDT' could mean 'Cuba Daylight Time' or 'Central Daylight
Time', for example.

A couple results of this:

* the series must advance a TimeWithZone, not a Time, to handle DST transitions
* the series needs a :zone options to specify what zone to use, it cannot be
  guessed

Advance to Spring DST
---------------------

Advancing to the spring DST transition (02:00) automatically causes times in a
DST-zone to flip to 03:00 -- the 02:00 time does not exist.  This causes
problems for advance, because chained advances through the missing time will
change the time (even if the period is 1 day, for example).  I'm not sure what
to do about this issue... some tests have been outlined and left commented out.

Calculating N Steps
--------------------

All of this has consequences for things like calculating n steps.  It is too
slow to advance through each step for large n; it's better to guess n, advance
to near-the-stop and check around there.  However there is no way to map a
logical period like {:months => 1} to a set number of seconds.  As best you can
do is use an average, then look down and up from wherever you end up.

As a result this changset adds some methods to multiply periods.