Rework Timeseries algorithm to preserve period
Chaining advance will, by the nature of Time, cause steps to fall out of
alignment in various circumstances.  Consider a series that starts at the
beginning of the month and chains advance(:months => 1).

    2010-01-01 00:00:00
    2010-02-01 00:00:00
    2010-03-01 00:00:00

This is the normal and expected behavior.  Now consider a series that starts at
the end of a 31-day month.

    2010-01-31 00:00:00
    2010-02-28 00:00:00
    2010-03-28 00:00:00

The day falls out of alignment because there is no way to logically advance to
2010-02-28 and preserve the information that this is a series starting on day
31.  I don't think there is a good way around that fact when chaining advances
-- if you advance to a date that doesn't exist (spring daylight saving 02:00 is
another example) then you *must* convert to some other date, and therefore the
intention of the series will always be lost.

Instead, calculate the time of each step by multiplying the period and
advancing from start.  That way each step remains true to the intent of the
series -- a number of periods relative to some start point vs an absolute
advancement between steps (which can only exist in cases where an 'absolute
advancement' actually exists -- like UTC and seconds).

    2010-01-31 00:00:00
    2010-02-28 00:00:00
    2010-03-31 00:00:00

See also https://github.com/rails/rails/issues/10938