Allow for period adjustment on stdin (-w)
This feature is useful when timeseries is the driver for realtime data
generation (or simulated realtime via a throttle).

I'm not quite sure how to test this given it's timing-sensitive nature.  I'm
also not sure how to expand it to -LR, allowing warping n_steps, stop_time, etc
given that will require having each respond mid-stream to a reconfig.