Rework timeseries interface
This major rework of the interface preserves most of the previous behavior but
refactors it into a more flexible set of options geared around input and output
modes.  These are breaking changes and will require a major version bump.

Notably:

The restart option -c has been removed.  Although this option worked for the
most part it never really could reliably restart because the recipient of the
output is the one that needs to dictate where a series should restart.

The pattern of -lower -UPPER where the lower option would be the infinite
series and the upper would be the constrained series has been replaced with
distinct input/output mode plus an option -m that toggles infinite output.
This reduced a great deal of duplication and freed up several letters for other
options.

A method implemented to do what a `Time.zone.strptime` method should do -- ie
`Time.zone.parse` but with an input format.  To be clear this method is a hack
and should probably exist in ActiveSupport.

Testing was generally expanded but more testing and cleanup is still needed.