Allow -d to specify per-feed attributes (removed dir support)
Previously the data resolution was for a particular feed value (which maybe
wasn't so clear) but frequently per-feed attributes are needed.  This
necessitated the addition of a nesting key to the data file because JSON will
not allow numeric keys.

There isn't a good generic solution for how to add this data for a directory,
without specifying another directory specifically for the feed attributes; all
reasonable file names (ie those without spaces) are accessible from the feed
values, so there is no room for per-feed files.  This being the case the
directory-style has been removed.  Maybe this is a pure good as the directory
style is hard to explain alongside the file solution.