Allow -cC to read header fields from input file
The data file can specify headers but doing so there is both a pain and
separates the headers from the data (where they are useful).  This makes it so
each data series can be fully identified in place.  Perhaps later it will be
better to remove the data file verison, usage will tell.