Update tests to reflect asymmetric MST/MDT transitions
Justification is in the comment, but to reiterate:

MST/MDT transitions are variable in the forward and back direction... as
justification if you are Jan 31 and you want to go forward 1 month then you end
up at Feb 28 or Feb 29, but if you go foward 2 months you end up at March 31.
Point being that if you can advance one digit in the time without changing any
of the others then you do it. It's only when you can't go there exactly that
you start munging other digits. Insofar as MST/MDT is a "digit" it should not
be munged unless needed. That means:

  MDT -> MDT -> MST  # forward
  MDT <- MST <- MST  # reverse